@@ -1,18 +1,16 @@
 package com.here.mobility.sdk.sampleapp.public_transport;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.here.mobility.sdk.demand.PublicTransportRouteLeg;
-import com.here.mobility.sdk.sampleapp.R;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Locale;
 import java.util.concurrent.TimeUnit;
 
 /**********************************************************
