@@ -1,21 +1,18 @@
 package com.here.mobility.sdk.sampleapp.rides;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.UiThread;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.annotation.UiThread;
+import androidx.recyclerview.widget.RecyclerView;
 import android.text.format.DateUtils;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import com.google.common.collect.Lists;
 import com.here.mobility.sdk.demand.PriceEstimate;
 import com.here.mobility.sdk.demand.Ride;
-import com.here.mobility.sdk.demand.RideStatusLog;
 import com.here.mobility.sdk.sampleapp.R;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
